// PAGridCustomizerCtrl/types.ts
export interface PAOneGridCustomizer {
  /**
  * Provides customizations for the grid.
  * This will include prop overrides, cell renderers and editors,
  * @notice If you just need to override few cell renderers or editors,
  * its better to use cellRendererOverrides, cellEditorOverrides props
  */
  gridCustomizer?: GridCustomizer;
  /** Provides overrides for built -in cell renderers */
  cellRendererOverrides?: CellRendererOverrides;
  /** Provides overrides for built -in cell editors */
  cellEditorOverrides?: CellEditorOverrides;
}

/**
 * Provide cell renderer overrides per column data type.
 */
export type CellRendererOverrides = {
  [dataType in ColumnDataType]?: (props: CellRendererProps, rendererParams: GetRendererParams)
    => React.ReactElement | null | undefined;
};

/**
 * Provide cell editor overrides per column data type.
 */
export type CellEditorOverrides = {
  [dataType in ColumnDataType]?: (defaultProps: CellEditorProps, rendererParams: GetEditorParams)
    => React.ReactElement | null | undefined;
};

export interface GridCustomizer {
  /** Returns react element for the column headers */
  GetHeaderRenderer?(params: GetHeaderParams): React.ReactElement;
  /** Returns the cell renderer react element */
  GetCellRenderer?(params: GetRendererParams): React.ReactElement;
  /** Return renderer for a row in a loading state */
  GetLoadingRowRenderer?(): React.ReactElement;
  /** Returns the cell editor react element */
  GetCellEditor?(params: GetEditorParams): React.ReactElement;
  /** Returns the component and properties to use for empty grid overlay */
  GetNoRowsOverlayConfiguration?(): NoRowsOverlayConfig;
}

export interface CellRendererProps {
  /** Raw value of the cell */
  readonly value: unknown;
  /** True when grid is in RTL mode */
  readonly isRTLMode?: boolean;
  /** Callback indicating the grid cell has been clicked */
  readonly onCellClicked?: (event?: React.MouseEvent<HTMLElement, MouseEvent> | MouseEvent) => void;
  /** Returns the validation error for the cell */
  readonly validationError?: Error | null;
  /** True when cell should be right aligned */
  readonly isRightAligned?: boolean;
  /** Callback to pragmatically start editing the cell */
  readonly startEditing?: (editorInitValue?: unknown) => void;
  /** True when cell is on the last row of the grid */
  readonly isLastRow?: boolean;
  /** Grid row height in pixels */
  readonly rowHeight?: number;
  /** Cell column data type */
  readonly columnDataType?: ColumnDataType;
  /** Formatted value of the cell */
  readonly formattedValue?: string;
  /** HTML element containing the cell */
  readonly cellContainerElement?: HTMLElement;
  /** Cell error label Id */
  readonly cellErrorLabelId?: string;
  /** True when the cell column is editable */
  readonly columnEditable?: boolean;
}

export interface GetRendererParams {
  /** Column definitions for all visible columns in the grid */
  readonly colDefs: ColumnDefinition[];
  /** Cell column index */
  readonly columnIndex: number;
  /** Cell row data */
  readonly rowData?: RowData;
  /** True when tab navigation is allowed in the grid */
  readonly allowTabKeyNavigation?: boolean;
}

export interface CellEditorProps {
  /** True when cell is secured */
  readonly secured?: boolean;
  /** Raw value of the cell */
  readonly value: unknown;
  /** True when grid is in RTL mode */
  readonly isRTLMode?: boolean;
  /** Grid row height in pixels */
  readonly rowHeight?: number;
  /** True when the cell value is required to be set */
  readonly isRequired?: boolean;
  /** Passing the first character pressed starting the cell editing */
  readonly charPress?: string | null;
  /** Cell column data type */
  readonly columnDataType?: ColumnDataType;
  /** Callback to notify the grid when value of the cell editor is changed */
  readonly onChange: (newValue: unknown) => void;
}

export interface GetEditorParams {
  /** Column definitions for all visible columns in the grid */
  readonly colDefs: ColumnDefinition[];
  /** Cell column index */
  readonly columnIndex: number;
  /** Callback to notify the grid when value of the cell editor is changed */
  readonly onCellValueChanged: (newValue: unknown) => void;
  /** Cell row data */
  readonly rowData?: RowData;
  /** Callback to programmatically stop the cell editing */
  readonly stopEditing: (cancel?: boolean) => void;
}

export interface GetHeaderParams {
  /** Column definitions for all visible columns in the grid */
  readonly colDefs: ColumnDefinition[];
  /** Cell column index */
  readonly columnIndex: number;
  /** True when header cell is the first column header */
  readonly isFirstVisualColumn?: boolean;
  /** True when header cell is the last column header */
  readonly isLastVisualColumn?: boolean;
  /** Cell row data */
  readonly rowData?: RowData;
  /** True when grid is in RTL mode */
  readonly isRTLMode?: boolean;
  /** True when tab navigation is allowed in the grid */
  readonly allowTabKeyNavigation?: boolean;
}

export interface NoRowsOverlayConfig {
  /** Component to render when no rows are present */
  component: React.ComponentClass | undefined;
  /** Properties to pass to the no row component */
  props: unknown | undefined;
}
// ...
export type ColumnDataType =
  | 'Text'
  | 'Email'
  | 'Phone'
  | 'Ticker'
  | 'URL'
  | 'TextArea'
  | 'Lookup'
  | 'Customer'
  | 'Owner'
  | 'MultiSelectPicklist'
  | 'OptionSet'
  | 'TwoOptions'
  | 'Duration'
  | 'Language'
  | 'Multiple'
  | 'TimeZone'
  | 'Integer'
  | 'Currency'
  | 'Decimal'
  | 'FloatingPoint'
  | 'AutoNumber'
  | 'DateOnly'
  | 'DateAndTime'
  | 'Image'
  | 'File'
  | 'Persona'
  | 'RichText'
  | 'UniqueIdentifier';

  export interface SettingModel {
    entity: string;
    attributes: string[];
    dropDownOptions?: any[];
    workOrderInfo: any;
    context?: any;
  }